+++ docs/changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BanCode` scanner to detect and block code snippets in the prompt.
 - Benchmarks on the AMD CPU.
 - `API` has a new endpoint `POST /scan/prompt` to scan the prompt without sanitizing it. It is faster than the `POST /analyze/scan` endpoint.
+- Example of running [LLM Guard with ChatGPT streaming mode](./tutorials/openai.md) enabled.
 
 ### Fixed
 - `InvisibleText` scanner to allow control characters like `\n`, `\t`, etc.